requirements.txt
# Core dependencies
streamlit>=1.32.0
pandas>=2.1.0
numpy>=1.24.3
openpyxl>=3.1.2

# Bioinformatics
biopython>=1.81
biopandas>=0.4.1

# Data visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.11.3
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Molecular visualization
py3Dmol>=2.0.0
stmol>=0.0.6
ipywidgets>=8.0.0
ipykernel>=6.0.0
Pillow>=10.0.0
cairosvg>=2.7.0

# Utilities
requests>=2.31.0
tqdm>=4.66.1
loguru>=0.7.2
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Structure prediction (placeholder implementation; external tools not required)

# Docking

# Additional data processing
pdb2pqr>=3.5.2
